requirements.txt
requests>=2.31.0
selenium>=4.16.0
pyautogui>=0.9.54
python-dotenv>=1.0.0